% Generated by build script
\name{ebi_features}
\alias{ebi_features}
\title{Get protein features/annotations from UniProt}
\usage{
ebi_features()
}
\arguments{
\item{accession}{UniProt accession ID (e.g. "P38398")}
}
\value{tibble: accession, type, description, begin, end, evidence}
\description{Get protein features/annotations from UniProt}
