% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebi.R
\name{ebi_search}
\alias{ebi_search}
\title{Search UniProt proteins}
\usage{
ebi_search(query, size = 10, offset = 0)
}
\arguments{
\item{query}{Search query string (gene name, protein name, etc.)}

\item{size}{Number of results (default 10, max 100)}

\item{offset}{Starting offset for pagination (default 0)}
}
\value{
tibble: accession, id, gene, organism, protein_name, existence,
sequence_length
}
\description{
Search UniProt proteins
}
